pyproject.toml
[project]
name = "repro-azure-dependency"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.23.1",
    "azure-storage-blob>=12.26.0",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.69",
    "langchain-unstructured>=0.1.6",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "unstructured[pdf]>=0.18.9",
]
